MdLinks function show links and resolve promises
